v1.3.0
Big changes in this push! Version 1.3.0 introduces a couple of new
features, provides several QoL changes, and a few optimization
changes under the hood. Some of note include:

New features
- New CLI detection logic: download your bw.exe CLI file (please use the latest one), and go into your plugin settings to point to it and then, if it doesn't detect the location in your PATH, just click the text to have it do it for you!
- Added the ability to clear your clipboard automatically based on your preference.
- Added sync functionality - now you can sync your vault on demand which includes the process of updating your icons.

Fixes/Improvements
- Fixed a bug with the URI window not popping up after several changes.
- Fixed the icon querying and caching mechanisms. Now the query occurs after unlock and manual sync asychnrounosly (with the unlock process) and will no longer hang up the Flow Launcher UI causing incorrect selection references.
- Removed the /unlock command from appearing after your vault was already unlocked.
- Fixed issue with requiring a restart of the Flow Launcher plugin to commit api credentials.
- Fixed issue where I had to do funky workarounds to get the plugin to work without sending the api credentials by having the user sign in in advance of using the plugin so that a session key could be created and managed due to a bug in previous versions of the CLI not actually allowing you to properly sign in using the apikey parameter.

There's probably a few more I'm not thinking of, and maybe it's not as big of an update overall, but I've been working on this a while
trying to get all the kinks out for the basic functionality and focusing hard on minimizing any actual interaction with the CLI by the user
which I think I have effectively done now. In my testing, I did not once actually have to open up the command line/PowerShell which will hopefully be a big improvement in the user experience.

Hopefully this will make the experience much more seamless. Next updates I'll try to include configuration settings to use your own hosted bitwarden/vaultwarden server so that it will hopefully completely erase the need to use the terminal ever.

Thanks all for your patience 😊